requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic[email]==2.7.0
discord.py==2.3.2
aiohttp==3.9.5
